First shot at kind coercion
Make all HLSL variables reference a global list of kinds,
make OperatorTypeSpec return a set of constraints on those kinds,
made the variable manager able to merge the kinds referenced
by variables to enforce those constraints.

Currently breaks when generating input/outputs such that X = X < (float)Y -
the rules state X should be both float because it's compared to a float,
and int because the output of FauxBoolean::LT is int.
If the kind of X is previously NUMERIC, the output won't be split off
into a new variable.