Separated VMScalarNameRef from TVM::ScalarDataRef, moved away from TypedVMRef
A ScalarNameRef is based on a VectorNameRef, which only has a base type
mask. This means one cannot perform argument-based type refinement directly,
and would need TypedVMRef. I couldn't remove TypedVMRef as I wanted to before,
but this does lessen it and remove it from LegacyOutcome and Outcome.
Originally I wanted to remove TypedVMRef because it interferes with
scalarizing - (TypedVMRef, VectorComponent) -> TypedVMRef(Name, Component)
is cumbersome - but removing it from the Outcomes is enough.